requirements.txt
requests==2.28.1
psycopg2-binary==2.9.4
